#55: Refactored the devices module. (#57)
